+++ Src/Roflcopter.Plugin/TodoItems/TodoItemsCountDefinition.cs
@@ -1,0 +1,14 @@
+namespace Roflcopter.Plugin.TodoItems
+{
+    public class TodoItemsCountDefinition
+    {
+        public TodoItemsCountDefinition(string name)
+        {
+            Name = name;
+        }
+
+        public string Name { get; }
+
+        public override string ToString() => Name;
+    }
+}