+++ Src/TestData/Roflcopter.Sample/Roflcopter.Sample/SomeClass.cs
@@ -7,6 +7,14 @@
         public void Method()
         {
             // todo
+
+
+            // TODO 1
+            // TODO 2
+            // TODO 3
+
+            // BUG 1
+            // BUG 2
         }
 
         public class NestedCass